Add, Edit, and Delete Buttons are working as intended
The buttons are working fine, except for some specific cases that i'll handle later.
The MediaList gets updated automatically when the user adds a media, the same goes for Edit and Delete; the last one is still inconsistent however.
The DescriptionTable gets updated when the user selects, adds, edits, or deletes a media from the list.
Finally, the media description is too long for the table cells and font size, so i'll have to add a ToolTipText to show the full content of each cell on hover.